requirement.txt
openai>=1.14.0
requests>=2.31.0
Pillow>=10.0.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.96.0
